add SerializeWithSurrogate trait
